paper-bill-detector/requirements.txt
fastapi
uvicorn
ultralytics
pillow
numpy
opencv-python-headless
python-multipart
gunicorn